Moved functions to projectedPotential.cpp
